* Implement methods for explicit casting values to generic type T
* Fix benchmarks to avoid overflowing of _counter and to make GC harder

+++ src/NoBox.Benchmarks/Generators/ShortBoxedValuesGenerator.cs
@@ -4,7 +4,7 @@ namespace Pankraty.NoBox.Benchmarks.Generators
 {
     internal class ShortBoxedValuesGenerator
     {
-        private int _counter;
+        private long _counter;
 
         public object GetNext()
         {
